+++ dbms/src/Storages/Transaction/TiDBSchemaSyncer.h
@@ -18,13 +18,14 @@
 #include <Common/TiFlashMetrics.h>
 #include <Debug/MockSchemaGetter.h>
 #include <Debug/MockSchemaNameMapper.h>
-#include <Storages/Transaction/SchemaBuilder.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <Storages/Transaction/TiDB.h>
 #include <pingcap/kv/Cluster.h>
 #include <pingcap/kv/Snapshot.h>
 
 #include <ext/scope_guard.h>
+
+#include <TIDB/Schema/SchemaBuilder.h>
 
 namespace DB
 {
